+++ lib/events/other/another_roll.rb
@@ -7,15 +7,21 @@ module TohsakaBot
       rate_limiter.bucket :roll, delay: 10
       extend Discordrb::EventContainer
 
-      BOT.button(custom_id: /^roll:\d/) do |event|
+      BOT.button(custom_id: /^roll:(\d|id)/) do |event|
         next if event.user.bot_account
 
         user_id = event.user.id.to_i
         name = BOT.member(event.server, user_id).display_name.sanitize_string
 
-        roll_size = event.custom_id.split(':').last.to_i
-        number = rand(0..("9" * roll_size).to_i).to_s
-        content = "🎲 **#{number.rjust(roll_size, '0')}** `#{name}`"
+        event_id = event.custom_id.split(':').last
+        number = '0'
+        if /\d/.match(event_id)
+          roll_size = event_id.to_i
+          number = rand(0..("9" * roll_size).to_i).to_s
+          content = "🎲 **#{number.rjust(roll_size, '0')}** `#{name}`"
+        else
+          content = "#{CommandLogic::Roll.parse_msg_id(event.interaction.id)} `#{name}`"
+        end
 
         event.defer_update
         BOT.send_message(event.channel, content, false, nil, nil, false, event.message.message_reference)